konsul-pajak-preprocessing-data/pyproject.toml
[project]
name = "konsul-pajak-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Novandra Anugrah",email = "156439898+nvn01@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langchain (>=1.0.4,<2.0.0)",
    "langchain-openai (>=1.0.2,<2.0.0)",
    "openai (>=2.7.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "pypdf (>=6.1.3,<7.0.0)",
    "chromadb (>=1.3.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
